+++ packages/vscode/src/state.ts
@@ -7,7 +7,7 @@ import { Utils } from "vscode-uri"
 import { listFiles, saveAllTextDocuments } from "./fs"
 import { parseAnnotations } from "../../core/src/annotations"
 import { Project, PromptScriptRunOptions } from "../../core/src/server/messages"
-import { JSONLineCache } from "../../core/src/cache"
+import { JSONLineCache } from "../../core/src/jsonlinecache"
 import { ChatCompletionsProgressReport } from "../../core/src/chattypes"
 import { fixCustomPrompts, fixPromptDefinitions } from "../../core/src/scripts"
 import { logMeasure } from "../../core/src/perf"
